Initial driver.py refactoring into several modules
