+++ tests/test_campaign.py
@@ -10,16 +10,14 @@ import yaml
 
 from hpcbench.campaign import default_campaign, pip_installer_url
 from hpcbench.cli import bensh
-from hpcbench.driver import (
-    BenchmarkCategoryDriver,
+from hpcbench.driver.executor import ExecutionDriver
+from hpcbench.driver.campaign import CampaignDriver, HostDriver, BenchmarkTagDriver
+from hpcbench.driver.benchmark import (
     BenchmarkDriver,
-    BenchmarkTagDriver,
-    CampaignDriver,
-    ConstraintTag,
-    ExecutionDriver,
+    BenchmarkCategoryDriver,
     FixedAttempts,
-    HostDriver,
 )
+from hpcbench.driver.base import ConstraintTag
 from hpcbench.toolbox.contextlib_ext import modified_environ
